+++ src/fair_and_square/src/library/levels/MazeLevel.hh
@@ -15,11 +15,11 @@ namespace octopus
 } // namespace octopus
 
 
-std::list<octopus::Steppable *> MazeLevelSteps(octopus::Library &lib_p, size_t number_p);
+std::list<octopus::Steppable *> MazeLevelSteps(octopus::Library &lib_p, uint32_t number_p);
 std::list<octopus::Command *> MazeLevelCommands(octopus::Library &lib_p);
 
 /// @brief write header for classic arena level
-void writeMazeLevelHeader(std::ofstream &file_p, size_t number_p);
+void writeMazeLevelHeader(std::ofstream &file_p, uint32_t number_p);
 /// @brief read header for classic arena level and return a pair of steppable and command
 std::pair<std::list<octopus::Steppable *>, std::list<octopus::Command *> > readMazeLevelHeader(octopus::Library &lib_p, std::ifstream &file_p);
